added spidey
added the spider to the main along with all of the components it will need, added the categories for personal info as items that the spider can use and send. as of right now it's hardcoded to scrape whitepages for my info but can't connect because I need to get a proxy and connect it but I'm waiting for my paycheck lmao.
TODO:
connect the search class to the spider so we can send potential links to the spider and input them as allowed domains and train the spider off of them to collect new info.