Add files via upload
Alright. This should probably be functional by itself (there's more unnecessary files). Anyways, I changed some of the textfields so that it's more sensitive to syntax. I additionally changed part of the address' fields and hometown's field entirely; it now asks for state instead of area, and zipcode instead of town. These changes allow for a more 'versatile' search query; the software is now vaguely capable of searching through an inputted address.